requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic
apscheduler
torch
transformers
datasets
beautifulsoup4
peft
accelerate
py_vncorenlp
pydantic_settings
motor
qdrant_client
